+++ src/main/java/domain/janggi/Score.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.janggi;
 
 public record Score(
         float value
