+++ src/main/java/domain/piece/Chariot.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.Team;
+import domain.janggi.Team;
 import domain.board.Offset;
 import domain.board.movement.Movement;
 import java.util.List;
