+++ src/main/java/domain/piece/Horse.java
@@ -9,7 +9,7 @@ import static domain.board.Offset.RIGHT_DOWN;
 import static domain.board.Offset.RIGHT_UP;
 import static domain.board.Offset.UP;
 
-import domain.Team;
+import domain.janggi.Team;
 import domain.board.Offset;
 import domain.board.movement.Movement;
 import java.util.List;
