+++ src/main/java/domain/piece/Piece.java
@@ -1,7 +1,7 @@
 package domain.piece;
 
-import domain.Score;
-import domain.Team;
+import domain.janggi.Score;
+import domain.janggi.Team;
 import domain.board.BoardPosition;
 import domain.board.Offset;
 import domain.board.movement.Movement;
