+++ src/main/java/manager/JanggiManager.java
@@ -3,11 +3,11 @@ package manager;
 import dao.ConnectionProvider;
 import dao.JanggiDao;
 import dao.PiecePositionDao;
-import domain.Janggi;
-import domain.JanggiStatus;
-import domain.Score;
-import domain.Team;
-import domain.Turn;
+import domain.janggi.Janggi;
+import domain.janggi.JanggiStatus;
+import domain.janggi.Score;
+import domain.janggi.Team;
+import domain.janggi.Turn;
 import domain.board.Board;
 import domain.board.BoardPosition;
 import domain.piece.Piece;
