+++ src/test/java/domain/JanggiTest.java
@@ -7,6 +7,10 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import domain.board.Board;
 import domain.board.BoardPosition;
 import domain.board.InitialBoardFixture;
+import domain.janggi.Janggi;
+import domain.janggi.Score;
+import domain.janggi.Team;
+import domain.janggi.Turn;
 import domain.piece.Cannon;
 import domain.piece.Chariot;
 import domain.piece.Elephant;
